Move Dockerfiles to frontend root to preserve local dev functionality
The command in the Readme (to run app locally) mounts the src directory
making it possible for local development changes to be reflected in the
container without restarting/rebeuilding it each time. A previous change
moved the Dockerfiles into a subdirectory which broke this functionality.

Additionally, with the Dockerfles in the subdirectory, the .dockerignore
wasn't being picked up at build time.

Also, at build time, with the files in a subdir, the context was incorrect.
Certain commands referenced files in the same directory as the Dockerfile
and this broke when the Dockerfile changed to a sub dir